requirements.txt
pytest
requests
python-dotenv
boto3
botocore